Move term database sygus to quantifiers engine, move some attributes to quantifiers attributes header.
